Get permissions in uninstall

+++ InstallerEventSourceCA/CustomAction.cs
@@ -15,9 +15,11 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 using System.Diagnostics;
+using System.Security.AccessControl;
 using System.Security.Principal;
 using System.ServiceProcess;
 using WixToolset.Dtf.WindowsInstaller;
+using static Vanara.PInvoke.AdvApi32;
 
 namespace InstallerEventSourceCA
 {
@@ -39,6 +41,7 @@ namespace InstallerEventSourceCA
                 switch (action)
                 {
                     case "Install":
+                    {
                         session.Log("Performing 'Install' action");
 
                         if (EventLog.SourceExists(appName))
@@ -53,11 +56,22 @@ namespace InstallerEventSourceCA
                         EventLog.CreateEventSource(appName, "Application");
 
                         break;
+                    }
                     case "UninstallInit":
+                    {
                         NTAccount serviceAccount = new("NT SERVICE", serviceName);
-                        SecurityIdentifier serviceAccountId = (SecurityIdentifier)
-                            serviceAccount.Translate(typeof(SecurityIdentifier));
-                        string sid = serviceAccountId.Value;
+                        string sid;
+                        try
+                        {
+                            SecurityIdentifier serviceAccountId = (SecurityIdentifier)
+                                serviceAccount.Translate(typeof(SecurityIdentifier));
+                            sid = serviceAccountId.Value;
+                        }
+                        catch (IdentityNotMappedException)
+                        {
+                            session.Log("Service is not installed");
+                            return ActionResult.Failure;
+                        }
 
                         string[] services =
                         [
@@ -65,13 +79,25 @@ namespace InstallerEventSourceCA
                             .. ServiceController.GetDevices().Select(s => s.ServiceName),
                         ];
 
+                        using SafeSC_HANDLE scm = OpenSCManager(null, null, ScManagerAccessTypes.SC_MANAGER_CONNECT);
+                        if (scm.IsInvalid)
+                        {
+                            session.Log("Could not connect to service manager");
+                            return ActionResult.Failure;
+                        }
+
                         foreach (string service in services)
                         {
-                            ServiceController serviceController = new(service);
+                            CommonSecurityDescriptor permissions = GetServiceSecurityDescriptor(service, scm);
+                            session.Log(
+                                $"Permissions for service {service}: {permissions.GetSddlForm(AccessControlSections.All)}"
+                            );
                         }
 
                         break;
+                    }
                     case "UninstallFinalize":
+                    {
                         session.Log("Performing 'UninstallFinalize' action");
 
                         if (!EventLog.SourceExists(appName))
@@ -102,13 +128,23 @@ namespace InstallerEventSourceCA
                             );
                             session.Log($"Deleting local files from {localFilesPath}");
                             DirectoryInfo localFiles = new(localFilesPath);
-                            localFiles.Delete(true);
+                            if (!localFiles.Exists)
+                            {
+                                session.Log("Directory does not exist. Already deleted");
+                            }
+                            else
+                            {
+                                localFiles.Delete(true);
+                            }
                         }
 
                         break;
+                    }
                     default:
+                    {
                         session.Log($"Invalid action: '{action}'");
                         return ActionResult.Failure;
+                    }
                 }
 
                 session.Log("End InstallerEventSourceCA");
@@ -120,6 +156,63 @@ namespace InstallerEventSourceCA
             }
 
             return ActionResult.Success;
+        }
+
+        private static CommonSecurityDescriptor GetServiceSecurityDescriptor(string serviceName, SafeSC_HANDLE scm)
+        {
+            using var service = OpenService(scm, serviceName, (ServiceAccessTypes)ServiceAccessRights.READ_CONTROL);
+            if (service.IsInvalid)
+            {
+                throw new Exception("Could not open service with READ_CONTROL");
+            }
+
+            if (
+                !QueryServiceObjectSecurity(
+                    service,
+                    Vanara.PInvoke.SECURITY_INFORMATION.DACL_SECURITY_INFORMATION,
+                    out var res
+                )
+            )
+            {
+                throw new Exception("Could not query service object security");
+            }
+
+            using var securityObject = res;
+
+            string sddl = ConvertSecurityDescriptorToStringSecurityDescriptor(
+                securityObject,
+                Vanara.PInvoke.SECURITY_INFORMATION.DACL_SECURITY_INFORMATION
+            );
+
+            return new(false, false, sddl);
+        }
+
+        private static void SetServiceSecurityDescriptor(
+            string serviceName,
+            SafeSC_HANDLE scm,
+            CommonSecurityDescriptor securityDescriptor
+        )
+        {
+            using var service = OpenService(scm, serviceName, (ServiceAccessTypes)ServiceAccessRights.WRITE_DAC);
+            if (service.IsInvalid)
+            {
+                throw new Exception("Could not open service with WRITE_DAC");
+            }
+
+            using var securityObject = ConvertStringSecurityDescriptorToSecurityDescriptor(
+                securityDescriptor.GetSddlForm(AccessControlSections.All)
+            );
+
+            if (
+                !SetServiceObjectSecurity(
+                    service,
+                    Vanara.PInvoke.SECURITY_INFORMATION.DACL_SECURITY_INFORMATION,
+                    securityObject
+                )
+            )
+            {
+                throw new Exception("Could not set service object security");
+            }
         }
     }
 }
